+++ src/endata.c
@@ -4166,7 +4166,7 @@ char worldStringBuf[750];
 FILE *worldBufHandle;
 
 /* Graphics animation state (from BSS) */
-unsigned int *colorTablePtr;
+uint16 *colorTablePtr;
 char inputChanged;
 int colorAnimIdx;
 int colorAnimEnabled;
